Rename {WatchingNotSupported -> NonWatching}VirtualFileSystem
and some more renames.
